float software emulation working

+++ src/OpenCL/poisson_kernel_initial.cl
@@ -53,7 +53,7 @@
 
 __kernel __attribute__ ((reqd_work_group_size(1, 1, 1)))
 __kernel void ops_poisson_kernel_initial(
-		__global double* restrict arg0,
+		__global float* restrict arg0,
 		const int base0,
 		const int size0,
 		const int size1,
